Enable jedi autocompletion call tips
Argument lists will be added to the autocompletion API. There is a bug
where the call tips will not appear unless the user either enters one
argument followed by a comma, or after typing the left brace, deletes
the brace and re-enters it.

Create a addToCompletionAPI method in SciptEditor that will add entries
to the existing API. This saves us clearing and then adding everything
each time we update. Keep a dict of completions, since we cannot retrive
the entries from the API.