Initial work on absorption corrections
Created new algorithm LoadSampleEnvironmentComplex that takes an xml file
as input. The xml file describes a set of .stl files corresponding to the various
environment components and for each a translation\rotation indicating how the components
should be assembled. The xml file also describes the material of each components

Fixed problem in SampleEnvironment.cpp where exception was raised if any environment components
didn't intersect with the beam profile eg a ring outside of the profile
Also fixed problem in same unit with sampling - the loop to find a neutron path that intersected
with an environment component was looping through 5000 iterations per env component before moving
onto another environment component. The loop should be the other way round to avoid sampling bias
towards environment components with small overlap with beam profile